File Upload: Ensure empty filename is passed to the server
In MS Edge, if an empty filename ('') is passed to the FormData object,
the browser assumes a default filename depending on the value. To ensure
the same behavior for all the browsers, we introduce an empty filename
string for empty filenames, which will then be removed again in the
upload request handler.

242031, 251366, 254533

Change-Id: I71fc37e19178bb447add753c29b8b07c2d2296e5
Reviewed-on: https://git.eclipse.org/r/155693
Reviewed-by: Andre Wegmueller <awe@bsiag.com>
Tested-by: CI Bot