File Upload: Ensure empty filename is passed to the server
In MS Edge, if an empty filename ('') is passed to the FormData object,
the browser assumes a default filename depending on the value. To ensure
the same behavior for all the browsers, we introduce an empty filename
string for empty filenames, which will then be removed again in the
upload request handler.

242031, 251366, 254533

Change-Id: Ic320b5de59869b752eba1f9ad24614b81776b9fa
Signed-off-by: Thomas Siegrist <thomas.siegrist@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/155725
Tested-by: CI Bot
Reviewed-by: Andre Wegmueller <awe@bsiag.com>