Improve productive JS stack traces
If an error occurs on a productive system, it is very hard to analyze
it because the stack trace points to the minified code.
This is actually nothing new, but because of the babel transformation,
not even the method names are readable now.

To improve it, source maps are now generated in prod mode, too, but
with important differences to the ones generated in dev mode:
1. They are generated as separate files to not bloat the actual source
files -> they are only requested when the browser's dev tools are
open.
2. They contain no source code which is especially important in a
closed source project.
3. The sources array doesn't point to the real files to prevent
revealing the absolute paths (security).

Having productive source maps is only the half step. They need to
be applied before the stack is sent to the server. For this, the
tool sourcemapped-stacktrace is used. The error handler now uses it
to transform the stack trace. Because it asynchronously downloads the
source map, the error handler is now async as well.

If an error occurs, the browser logs the error to the console.
In addition, the error handler applies the source map, analyzes the
error, shows a message box, logs the mapped error to the console
and sends it to the server (existing behavior, beside the mapping).
One could stop the browser from logging the error a second time
by returning true in the onerror handler, but it actually can be
beneficial because the links from the mapped stack are not clickable
in every browser (e.g. firefox).

There is one drawback: it takes some time until the error message is
shown, before it was shown instantly.

261160

Change-Id: Iabdfb08ad9ce7f882ff44d3fb272bf4ed3af7388
Reviewed-on: https://git.eclipse.org/r/155774
Tested-by: Scout Bot <scout-bot@eclipse.org>
Tested-by: CI Bot
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>