JsonTable: ignore inconsistent/obsolete selection events from UI
When a form is stored and closed, it is possible that the table page is
automatically reloaded when the desktop is activated (via data change
listener). For a short period, the user may click another table row,
while the data is reloaded (busy indicator is not yet active). This can
lead to an inconsistent state between UI server and UI:
- UI is waiting for a reply to the "desktop is now in foreground" event.
- The row selection event is queued, because another request is
  currently in progress.
- When the data has finished loading, all table rows are deleted and
  replaced with new table rows. All rowIds are new. The server also sent
  a new selected row, which is selected by the UI.
- The queued selection event is now sent to the UI server, but it still
  contains an old rowId. The UI server correctly filters this unknown
  rowId.
- Because no row remains, the selection is cleared on the UI server.
  The event filter prevents this new state to be propagated to the UI.
  Therefore, the UI now has a different state (1 selected row) as the
  UI server (0 selected rows).

Fix: Ignore events that contain _only_ invalid rowIds.

251131

Change-Id: I401be7ac1a07bc5d28e301b898060716a29be5e8