Support file with variable for use in config.properties
The scout config loader now supports importing an additional file
defining only variables.

This feature is helpful when launching applications in development mode
in an IDE such as eclipse and IntelliJ using a pre-defined
config.properties.

The variable file only contains variables and config property values but
is itself not a config.property file. Thus its values are not checked in
any of the ConfigurationValidator

A variable file is passed to the java process using the system property
-Dscout.env=file:/path/to/my/launch.properties

Signed-off-by: Ivan Motsch <ivan.motsch@bsiag.com>
Change-Id: I2d67b26a5ae641ed271fd0a6d59f93fe204eacec
Reviewed-on: https://git.eclipse.org/r/156093
Tested-by: CI Bot