Table: prevent exception when row is expanded and updated
When clicking the control to expand a row, a selection and
expansion event is sent to the server. The selection on mouse down,
expand on mouse up.
If the selection handler on server side updates the row, the current
state of the row is sent to the client, which contains the expanded
state. At that time, it was false. Because updateRows() does not
consider the expanded property but the model state was changed,
the ui was in an inconsistent state and the next click could provoke
an error.

Fixed by making sure updateRows() also rerenders the rows if expanded
state changed, even though there is actually a dedicated expansion event.
Unfortunately, this does not solve it completely. If the user releases
the mouse button very fast, the selection event will be processed before
the expansion event but the update event won't be returned before the expansion
event was sent -> collapsing won't work because the old state was applied
and the new event not sent to the ui again (-> row is collapsed on server
but not on ui).
Fixed by handling the expansion also on mouse down, as it is done for
Tree. This makes sure the events will be sent together.
Theoretically it could still happen if the selection handler on the
server would break the execution using waitFor... The only solution
we could think of would be to remove the filters so that the server
always confirms, but that's not worth it.

259808

Change-Id: I18b56fe97ac28d4e930c5131d42dc60b5adbe22c
Reviewed-on: https://git.eclipse.org/r/155935
Tested-by: CI Bot
Reviewed-by: Andreas Hoegger <Andreas.Hoegger@bsiag.com>