Tiles sometimes invisible when filter changes in an invisible container
Use case:
Tile accordion in a scrollable group box.
Tile accordion will get invisible if all tiles are filtered out.
If the filter changes to accept the tiles again, the tile
accordion will get visible. Unfortunately, the tiles are still invisible.
Reason: the virtual scrolling logic does not render tile grids outside
of the viewport. In that case the tile grid is not outside of the
view port, it is just invisible -> improve view port check.

261405