allow IAccessControllers to request update of the Subject
If an IAccessController decided to create a new Subject the UiSession
was not aware of this happening and created a RunContext with the old
Subject that was still saved on the IClientSession.
A mechanism has been added to allow the IAccessController to inform the
UiSession that a new Subject was created and should be updated on the
IClientSession.