TileGrid: ensure inserted tiles are always visible
It may happen that inserted tiles still have the class invisible
after they are inserted.

This can happen if the tiles are inserted while the group is expanding.
In that case the group validates the layout tree before the class
is added in _renderInsertTiles and not again afterwards -> move the
invalidation to _renderInsertTiles to guarantee it is always invalidated
after the class was added.

259762