Responsive: don't move label from placeholder to top
Since commit 38d765e0ec8820a39028f4dc108962b6374d9cd7, placeholder
fields can have a top label. This can break the responsive mode.

Example:
A field with inline label and a placeholder field on the same line ->
placeholder field will increase the line height and stretch the
field with the inline label
It should be safe to ignore placeholder fields in the responsive handler
when moving the label to top because they did not have that capability
at all previously.