Table: collapsing not possible when child rows are updated
If a child row is updated while the parent row is being collapsed,
the child row won't be removed and still visible -> collapsing
does not work.

If the child row is updated while the parent row is being expanded,
the animation would stop immediately -> expanding works but without
animation.

The first issue can be fixed easily, the second case can be fixed for
most
cases. Sometimes there won't be an animation, sometimes the animation
will start twice (depending on the network latency, if the row update
happens in a selection handler on the server).

254527

Change-Id: Ia690f1a0248a81aae658797ef61a7f17b7df0696
Reviewed-on: https://git.eclipse.org/r/155936
Tested-by: CI Bot
Reviewed-by: Andreas Hoegger <Andreas.Hoegger@bsiag.com>