Wizard: improve paddings of wizard steps
If opened in a view, the wizards steps bar has a bottom border that
is drawn to the very left and right. The menu bar also has a bottom
border but a shorter one. So if there is a menu bar and a wizard steps
bar, there are two lines with different lengths.

Also, if the wizard is opened in a dialog, the wizard steps have a large
padding on the left and right which is not necessary.

-> Use the same paddings everywhere and don't draw the line to the very
left and right. If needed for special cases, custom rules can always be
added.

261541