Improved handling for undo/redo events in SmartField
All browsers handle undo/redo events differently. This solution is based
on the input event which should work for all browsers. We simply open
the proposal popup again, if the user has not typed anything before.
This happens for instance when an undo-event affects a SmartField that
is currently not focused.

Since the SmartField cannot wait for a lookup-call when the user clicks
somewhere else, the value may still be incorrect in cases where the
lookup-call does not return before the user clicks outside of the
SmartField.

246765