Wizard: Close the wizard when its container form is disposed.
When the container form of the wizard is being disposed, its wizard was
not closed. Usually the forms of the wizard steps are not managed by the
container form and will therefore be disposed once the wizard is closed.
The wizard is no Widget and will not be closed while the session is
destroyed, leaving some step forms undisposed.

Added a listener to the wizard to close the wizard once the container
form is disposed.

Added new property PROP_MANAGED_BY_CONTAINER_FORM, which is true by
default, to disable the automatic closing of the wizard.

260397

Change-Id: Ibea8341df2a3f3735a8354851511945f37d82dfc
Signed-off-by: Thomas Siegrist <thomas.siegrist@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/156715
Reviewed-by: Arthur van Dorp <Arthur.vanDorp@bsi-software.com>
Tested-by: Arthur van Dorp <Arthur.vanDorp@bsi-software.com>