Smart Field: prevent loading data twice with "hierarchical = true"
When loading the lookup rows in a hierarchical smart field, the same
lookup call is sometimes called again. The problem is caused by
HierarchicalLookupResultBuilder. When it processed the rows, it needs to
resolve the parent row. Depending on the order of rows, the parent row
might not be in the key cache, because this cache was filled while
processing the rows.

Fix: first put all received rows in the key cache, then process them.

Change-Id: I07423ace291496e377a6acfbf308d120ec7e76ec
Reviewed-on: https://git.eclipse.org/r/156857
Tested-by: CI Bot
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>