Fix csv import to keep all sample in batched load
When a csv is loaded in batches and the stream reader is kept open, the
ready will loose a sample per batch. The reason is, that the loop
first reads a line and then the check is done if the batch has already
read completely. This results in a skipped sample when the next batch
is processed.

Signed-off-by: Nino Weingart <nino.weingart@bsi-software.com>
Change-Id: Ia9a8c6a729aef406cb22bd717eccb87a98fe84f4
Reviewed-on: https://git.eclipse.org/r/156875
Tested-by: CI Bot
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>