NLS: change default value of file watcher to enabled and fix tests
Most of the test will be running in development mode. Therefore the NLS
file watcher will be active. The AssertNoRunningJobsStatement ensures
that all the jobs started in a test are finished after the test. For the
nls file watcher job its ok, that it remains active. Therefore we
exclude it in the statement.

Change-Id: Ibab5fae17069f4427cdf83860fbcdc7ceb8944ba
Reviewed-on: https://git.eclipse.org/r/156873
Tested-by: CI Bot
Reviewed-by: Thomas Siegrist <thomas.siegrist@bsi-software.com>