Make double click listener less error prone
The double click listener has been added multiple times on children of
the tile grid when a developer added DOM elements to the tile that
accidentally have the class 'tile'. With this change we only add the
listener on immediate children of the grid.

A project added a table to a tile, this table had the 'tile' class too.
Thus the tile grid also tried to handle double clicks on that table,
resulting on errors on the UI server.

262146