StringField: missing border on read only multi line string fields
A disabled alternative styled multi line string field has a dashed
border around the whole field. A read only one only has a bottom border
but the text inside the field is not aligned with the other fields.

-> Either remove the padding to align it with regular fields, or add
the border around the whole field. Decided for the second option so that
it behaves the same like a regular disabled alternative styled field.
If there is only a bottom border it can look a little strange if the
field is big with little content.

239815