Added support for MultiStatus, migrated DateField to new concept
The FormField in Scout JS now works with a 'multi-status' in a similar
way the MultiStatus works in Scout Classic. Instead of having a single
error status at one time, the Status has children with different types:
ParsingFailedStatus, ValidationFailedStatus and other Status. These
status can co-exist, so when a programmer adds a functional error, this
error will not replace an existing parse or validate error.

In Scout JS the functions addErrorStatus and removeErrorStatus have
been added. These are required for functional errors, it is the
responsibility of the programmer, to add/remove this type of error.

The synchronisation between Scout JS and Classic has been improved.
The UI server now can flag certain Status objects with 'deletable=false'
which means Scout JS is not allowed to remove these Status objects.

Adjusted specs where required. All specs and selenium tests are green.

Change-Id: I31e41656fa62a9ef15f6cb4e032a8509ccf8a886