Add direct karma dependency to support IntelliJ's Karma plugin
The transitive dependency (from @eclipse-scout/cli) is obviously
not sufficient to activate the plugin.

Also add direct dependencies to the jasmine plugins we use and unexclude
them to make code completion work. It seems not to be necessary
for the other modules because they have a (maven) dependency to
eclipse-scout-core which seems to activate code completion for these
plugins as well.