NLS: Refactor nls file watcher to only one instance.
Ensure there is only one thread watching for nls file changes and not
one for every nls resource bundle cache as before. The resource bundle
caches now register a handler to be notified, when a relevant change
occurs.

The watching thread is changed from a scout job to a daemon thread, to
ensure the main loop will not be blocked by the watching thread.

Change-Id: If21081c27b3dc1aae9e334049f086eeeecb89ccd
Reviewed-on: https://git.eclipse.org/r/157593
Reviewed-by: Thomas Siegrist <thomas.siegrist@bsi-software.com>
Tested-by: CI Bot