BrowserField: location is loaded multiple times
Every time _renderLocation is called or the iframe or any parent appended
to the dom, the request is aborted and restarted again.
This can not be handled only in BrowserField js, because if a parent
calls addFieldContainer and the fieldContainer was already added,
the issue would happen as well.

The check for the sibling (next()) is actually only necessary because
some fields could have used the function to move the fieldContainer
to the end. To not break that code appendTo is called also if the
fieldContainer is not the last element.

260078