POC: add virtual focus
The idea is to defer the focus until the layouting process is done.
The big benefit is, that it does not matter when the widget.focus() or
focusManager.requestFocus() is called, the actual element.focus() call
will be executed at the right time -> no more unexpected scrolling
issues.

A first test was successful, but a lot of the Jasmine specs fail because
layout is not executed -> test adjustments are necessary. Also, there
some questions and tasks open, see TODOs in VirtualFocus.js

227039