IE: ugly flickering when changing table pages with many rows
Focus calls come too early when table is not correctly layouted yet.
-> if table is big ie tries to scroll which leads to ugly flickering
and moved desktop whenever a large page with table is clicked.

Delaying the focus call does not work unless all focus calls would be
delayed (which would be a major change, but could definitely be
an option one time).

227039