Exception on form close if a focused field wants to show a warning
Use case:
A field wants to display a warning using a validator.
If the field has the focus and the user clicks ok to close the dialog,
an exception occurs.

Reason:
When the field is removed, the focus manager sets the focus to another
field which triggers onBlur which triggers acceptInput. Accepting the
input triggers the validator which sets an error status on the field.
This error status should be displayed but it fails because the field
status is already removed -> Never try to show a tooltip if the field
status is removed.

250426
263437

Change-Id: I6a94434bc9e12f86ce4d7963ec6f5ae7973ca2c1
Signed-off-by: Roger Stebler <Roger.Stebler@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/157917
Tested-by: CI Bot
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>