AbstractDateColumn - Do not parse value when creating field from cell
When editing a table cell that already has an error status the table
column tries to parse the text before writing it to the generated edit
field.
However the parsing and validation of a date field is implemented in the
UI only. The AbstractDateField does not implement parseValueInternal.
This means the parsing of the cell value will fail and the cell will get
a "Not implemented" error status.

Solution is to not try to parse the value when editing a date column
cell. Instead the value is directly set as displayText.

241181

Change-Id: Id7137d1a66d8efa8de448d78dd0544f55f8a6c6b
Signed-off-by: Roger Stebler <Roger.Stebler@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/157980
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>
Tested-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>