Only attach close handler when menus are really rendered
The close handler logic for submenus should only be executed if the
menu's are really rendered.
There are cases when a submenus' selected property is true but the
submenu was never rendered. In those cases removeSubMenuItems (triggered
by parentMenu.setSelected(false)) will fail.
This fix prevents this corner case error.

261996

Change-Id: Id8868b4381f23c8b46dec687ea720050b49c376e
Signed-off-by: Ralf Muri <ralf.muri@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/158015
Tested-by: CI Bot
Reviewed-by: Roger Stebler <roger.stebler@bsi-software.com>
Signed-off-by: Ralf Muri <ralf.muri@bsi-software.com>