Add pnpm-workspace.yaml
Because we use a larger workspace not just containing Scout,
the Scout RT repository does not contain the pnpm-workspace.yaml.
When building, the jenkins job creates an empty file to make pnpm happy.
One drawback of the current approach is that manually running
mvn compile won't work due to the missing file.
Also, pnpm has to search the npm modules instead of just using the ones
defined in the pnpm-workspace.yaml.

-> Adding the file to GIT enables mvn compile and may even make the
build a little faster.

Adding lock file to git ignore because it would only be created
when this specific workspace is built, which is not the case on daily
work only when building -> file would never be up to date.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=560406