Fixed bug with wrong insert position of nodes in Tree
The bug occurred because the dimensions of the Tree and its nodes are
slightly different than in Scout 7.0. The property 'viewRangeSize' in
Scout 7.0 was never smaller than 4, in newer versions the value could
become 2 and as a consequence the insert position of the insertBatch
was not correct, leading to nodes inserted at the wrong position.

262890