TabBox: fix unexpected error: Could not resolve event targets
If a tab item that is made invisible using setVisibleGranted(false)
it must not be sent to UI. This already works. But if that tab
is selected using tabBox.setSelectedTab(tab), a json adapter would
be created accidentally. If this adapter or a child adapter receives
a model event, it will be added to the json response. But since the
adapter will never be sent to UI, the events will generate JS errors
saying the event target (adapter) cannot be resolved.

Fixed by making sure setSelectedTab does not accidentally create
a json adapter. The UI will now select no tab -> The developer should
make sure to not select invisible tabs.

263972