Table/Tree: Menu not working after selection change in disabled field
The TableContextMenu calculates its enabled state only on a TableEvent
but not when the table gets enabled/disabled. When the user changes
the row while the table is disabled the TableContextMenu gets disabled.
When the table gets enabled again the TableContextMenu stays disabled
and no action is executed when clicking the menu.

Solution is to also calculate the context menu's enabled state when the
table gets enabled/disabled.

The same happens with the TreeContextMenu in a Tree.

263734

Change-Id: I387fad28688fc035d9ebde54aa9487aaf21e0a30
Signed-off-by: Roger Stebler <Roger.Stebler@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/158581
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>
Tested-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>