Improved date prediction for patterns containing a weekday (EEE)
Before this bugfix, the prediction for dates with a weekday was not
accurate. For instance, when a user type 'Mo. 22' the prediction shown
in the background could change to 'So. 22.03.2020' and looked broken.

With this change the weekday is used to calculate the prediction date.
We look for the next date that matches the weekday _and_ the date. If
no suitable date is found, the prediction is aborted. Which means the
"broken" prediction state cannot occur anymore.

235975