Tree: Nodes may be inserted at wrong position on a filter change event
Since commit c8200dec the INSERT and DELETE events on a
NODE_FILTER_CHANGED event are grouped. First all INSERT events are
handled, then all DELETE events.
The model calculates the position for the nodes to be inserted and
stores them in the INSERT event's childNodeIndex.
However when the events are grouped and the INSERT events are all
handled first, these calculated childNodeIndex may not represent the
correct insert position on the UI because there still may exist nodes
that should be removed.

Solution is to handle all DELETE events first before handling the INSERT
events. DELETE events do not have this problem because they use the
node's ID instead of a position index to identify the node to be
deleted.

264258

Change-Id: I316adb67ff3a71764958504d5618f02a027cf155
Signed-off-by: Roger Stebler <Roger.Stebler@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/158744
Tested-by: CI Bot
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>