Avoid errors in Chrome while trying to read document height
When content is from a different domain (like YouTube embedded videos)
its not possible to access the contents to determine its height. Added
an additional check that notices that case and do not read the height
in that case. Basically this means useUiHeight does not work in that
case, but height is not set to 0 anymore, instead it keeps the size
it had before.

Additionally chrome needs the 'allow-presentation' sandbox permission
in order to display embedded YouTube videos without security errors.

263785