HtmlComponent: prevent wrong size measurements during CSS animation
When the layout of a component is validated while it or one of the
parents is currently running a CSS animation, the measured sizes might
be wrong, because the browser sometimes returns mid-animation values
(e.g. when using grow/shrink animations).

Example:
A popup with a list box is opened with a "fade in and grow" animation.
While the animation is running, the list box's lookup call has completed
the loading of the data and inserts the corresponding table rows. This
causes the table layout to be revalidated. Because the visible size on
screen is currently smaller than the defined CSS size, the table will
sizes incorrectly.

To fix this, the layout has to be postponed until the animation has
finished. There was already a similar check in HtmlComponent.validate-
Layout() that deals with invisible parents. A similar check was added
for animations. Because the component being layouted might be anywhere
inside the animated parent, all parents up to the DOM root have to be
checked. Additionally, the layout is also deferred when the current
element is animated. This requires an adjustment in Form.js,
because the 'animate-open' class in dialogs must only be added _after_
the layout has been computed initially (otherwise, the dialog position
would be wrong).

This change assumes that the animation is added using a CSS class that
starts with the string 'animate-', e.g. 'animate-open'. (Unfortunately
there is no generic way to determine if an animation is running on a DOM
element.)

Note:
This is a combined backport of the following commits in Scout 10.0:
ff90b2d00b5293618ac08f64b6677bae385837f9
12053cc71db2e702b767da6613922644dc017942

264084

Change-Id: I34b56963d17d14c41cf69f55e84ebc63c59cc647
Reviewed-on: https://git.eclipse.org/r/160113
Tested-by: CI Bot
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>