StringColumn: Changing displayText to null does not work
When a cell in a StringColumn has a value set and (only) the displayText
is updated to null the UI shows the column's value instead.
This case occurs for example when a VetoException is thrown in
execValidateValue of the StringColumn while the new value is null.

Solution is to use an empty string for the displayText if it's null.

265110

Change-Id: I3755cd6c3605945d57200ce14aa842285fc9c930
Signed-off-by: Roger Stebler <Roger.Stebler@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/159327
Reviewed-by: Andre Wegmueller <awe@bsiag.com>
Tested-by: Andre Wegmueller <awe@bsiag.com>