ajax: move error handling to AjaxCall
scout.ajax.call() returns a promise that in case of an HTTP error is
rejected with a single AjaxError argument. This conversion was moved
to AjaxCall.js, so all usages of AjaxCall can profit from this
conversion, even if they don't use the scout.ajax utility.

For required migration steps, see migration guide.