Fix various issues concerning popups, page reload & glasspanes
Popups don't have GlassPanes:
Popups are painted outside of the widget parent-child DOM hierarchy.
That's why the current GlassPanes have no effect on Popups.
-> Added Popups as glassPaneTargets

Layouting the MenuBar removes/renders the menuItems:
When removing a menuItem which is a ContextMenuPopup the glassPane div
is lost.
-> Check for active glassPanes in the postRender of the Popup and
rerender the glassPanes if necessary.

MessageBoxes/FileChooser loose the correct displayParent after refresh:
The displayParent property is not a model property but is only set by
the show/activate event. After a page reload with a open
MessageBox/FileChooser the displayParent is undefined. Set it according
the the Controller's displayParent as this is done with Forms as well.

265221

Change-Id: I4c35ad3c2b6a73573cc5c961bfeb61b698165c19
Signed-off-by: Ralf Muri <ralf.muri@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/159788
Tested-by: CI Bot
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>