AjaxCall: fix rejection reason
$.ajax() does not return a normal promise, therefore it cannot be
changed by simply adding a catch() clause to the promise chain.
To correctly convert the result of AjaxCall in the "reject" case, the
"this.result" value must be changed accordingly. Internally, AjaxCall.js
still needs to adhere to the jQuery API.

This is a fix for an error that was introduced with commit
d68421fe879932bedaecdecf2261d52d3d8fd9e7.