FormField: update menus when visibility changes
If the visibility of a menu changes the field status needs to be
updated.

As alternative to _updateMenus we could call setMenusVisible but it is
not clear what this property is used for. Originally it was added to
reflect the server status of all menus. But since there is a public
setter it could be used to toggle the visibility of the ellipsis menu.
We should either make it clear what the purpose is or remove the method
completely... In the meantime we use _updateMenus because it won't break
anything.

Also added methods to insert or delete menus more easily.

246873