SequenceBox: exception when menus are changed
If menus are deleted on the field the menus of the sequence box
should get visible. This did not work because calling setMenus() on the
sequence box deleted the menus. Reverting to the deleted menus threw
an exception.

Maybe it was a bad idea to implement this feature at all, it will
probably never work satisfying (e.g. maybe sometimes seq box menus
should be visible always. Or what should happen if all field
menus get invisible? ...)