Fix "Assertion error: Session is stopping"
When desktop closes, all forms will be closed. If one of the forms
is the display parent of another form the display parent will be set
to desktop. If the form is showing, hideForm() and showForm() are called
to properly reset the displayParent on the UI. This will throw an
exception because showing a form during session stop is now allowed.

Solution is to not try to hide and show the form when session is
stopping because it is not necessary anyway at that time.

265592

Change-Id: I7c7d2b55298b09b3f908e32a9ec428418f88f126
Reviewed-on: https://git.eclipse.org/r/159938
Tested-by: CI Bot
Reviewed-by: Ivan Motsch <ivan.motsch@bsiag.com>