TileTable: exception when context menu key stroke is pressed
ShowContextMenu determines pageX and pageY based on the selected
row / tile. This does only work if the data is rendered.

When the tile grid is active, the showContextMenu function of the
tile grid has to be used because it needs the selected tiles to
determine the position. But the menu items are only available on the
table -> pass the table menu items to the showContextMenu function
of the tile grid.

264372

Change-Id: I764eac25a5fc5206c5c97565b19061f92dca4350
Reviewed-on: https://git.eclipse.org/r/159942
Tested-by: CI Bot
Reviewed-by: Ralf Muri <ralf.muri@bsi-software.com>