Outline: don't navigate to top while deleting nodes
When all nodes are deleted, navigateToTop was executed for every
deleted node. This could even lead to exceptions (e.g. can only remove
nodes at the beginning or end of the existing range).

NavigateToTop should actually only be executed when the last node is
deleted, but this.nodes is set to [] before onNodeDeleted is called.

The solution is now to completely remove navigateToTop and fire a
selection event instead when the nodes are deleted. This already
happens for Table.js, TileGrid.js and as well in the Java code
(AbstractTree, etc.).
An alternative would have been to additionally listen for deleted
nodes in the DesktopBench and to update the OutlineContent when
the selected node is deleted, but since other widgets already fire
the selection even it is consistent to do so for the Tree as well.

265861

Change-Id: Ibe87ef40cbf11b7ce3216856640eb1e443aea102
Reviewed-on: https://git.eclipse.org/r/160101
Tested-by: CI Bot
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>