HtmlComponent: also delay layout if current element is being animated
This is an addendum to commit ff90b2d00b5293618ac08f64b6677bae385837f9.

In addition to the original change, the layout is also deferred when the
current element is being animated. This requires an adjustment in
Form.js,because the 'animate-open' class in dialogs must only be added
_after_ the layout has been computed initially (otherwise, the dialog
position would be wrong).

264084

Change-Id: I59e470b267bf6ffc616ab6b48b23a90e9b78ffd6
Reviewed-on: https://git.eclipse.org/r/160104
Tested-by: CI Bot
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>