Improved deep-link handling on startup
When the desktop is initialized and creates the first browser history
entry for the activated outline deep-link, we now suppress the deep-
link path in the URL of the browser. This gives the user the chance to
bookmark the initial startup URL (which triggers the startup bookmark)
whereas a bookmarked URL with a deep-link does _not_ trigger the startup
bookmark.

265433