Table: error when values of a smart column resolve immediately
The error occurred when rows are loaded into a table with a smart column
which will resolve the text of the cells later. But only if the first
rows did not contain any value so that the text was resolved
immediately. And only if there is a sort column active so that resolving
the null value would result in a row order changed event.

Fixed by making sure the update buffer will always process async.

Affects only Scout JS.

266095