TileGrid: error when context menu keystroke is used but tile not in view
When pressing the context menu key stroke the context menu should
be positioned on the selected tile. If the tile is not in view,
an exception occurs.
Context menu is now drawn in the top left corner of the tile grid
if the selected tile is not in view. Furthermore, the focused tile
is now used instead of the last tile in the selection. This is more
natural especially when multi selecting tiles (focused tile is the last
tile the user clicked).

Also removed hybrid-scrollable class when scrollable is toggled
because it was confusing when the container has this class but
is not scrollable at all.

264372