Layout: improve layout caching and remove unnecessary layout calls
If setSize is used with fractional values, the wrong size is cached
(without fractional values) so that the component will always be
invalidated and the layout done even if it would not be necessary.
Fixed by using the exact size when caching.
Also removed some unnecessary validateLayoutTree calls, the desktop
bench itself does it already, no need for the bench column to do
it as well.

The unnecessary layout invocations can even lead to exceptions in the
tree. Example: first node is selected, outline is scrolled so that
the first node is detached, outline is reloaded -> outline content
will be removed, invalidateLayoutTree called from the bench
but still the outline layout executed due to the not working cache.
TreeLayout calls renderViewPort which throws an exception because
deleting nodes is still in progress.

265861

Change-Id: I5ba24ce4efc73d946392b227f752a05c8ff339f9
Reviewed-on: https://git.eclipse.org/r/160604
Tested-by: CI Bot
Reviewed-by: Andre Wegmueller <awe@bsiag.com>