Widget: invalidate parent logical grid regardless if rendered
We must invalidate the parent logical grid regardless if the current
widget is rendered or not, otherwise the grid will be messed up.

Problem:
We have invisible fields in a box that is currently not rendered,
because the box will be displayed dynamically within a popup. If we set
these fields to be visible and open the popup again the newly visible
fields are not placed properly within the grid.

Cause:
The method to invalidate the parent logical grid only invalidate the
grid if the current widget is also rendered.

Signed-off-by: Torsten Trompler <Torsten.Trompler@bsi-software.com>
Change-Id: I6b24d6cbb2ddcc965719c88ca2265f22dabeddb5
Reviewed-on: https://git.eclipse.org/r/160636
Tested-by: CI Bot
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>