cli: process arguments to allow adjusting the build using cli
- The command line arguments are now passed to webpack and karma
when calling a scout script. This makes it possible to override
configuration settings using cli. Examples:
npm run build:dev -- --progress false
npm run build:ci -- --junitReporter.outputDir=./target/test-results

Currently supported arguments for the webpack build are:
mode, clean, progress, profile, resDirArray, stats

For karma, the arguments are passed to the karma runner which
just merges the arguments with the config. This means, all config
properties are supported but not all work because the props may
have the wrong type (e.g. auto-watch false won't work because it will
be a string, but no-auto-watch will work). We could optimize this
by defining the type for every karma property but this does not seem
to be necessary so far.

- The webpack log output can now be controlled using the stats property.
The property can either be set in the config or as command line
argument. See https://webpack.js.org/configuration/stats/ for details.

- The watch parameter has been renamed to clean because it does
not actually control the watching, it only controls cleaning the output.

- The env parameter passed to the config function previously was
process.env. This is different to the webpack-cli which just passes
args.env. Because people may already know the webpack-cli it could
be confusing if the env parameter contains something completely
different. It is still possible to access process.env directly in the
config file.

- Removed obsolete dependencies: it is not clear why webpack-cli,
webpack-dev-server and webpack-shell-plugin are listed as dependencies.
Everything seems to work just fine without them.

267177

Change-Id: I4c6f8504647372b3687ccf8aa685cbba145de013
Reviewed-on: https://git.eclipse.org/r/160720
Tested-by: CI Bot
Reviewed-by: Matthias Villiger <mvi@bsi-software.com>
Reviewed-by: Daniel Schmid <daniel.schmid@bsiag.com>