Releng: consider pnpm-workspace.yaml when looking for workspace modules
The releng script currently searches all package.json files starting
from the pnpm-workspace.yaml. This does not work if pnpm-workspace.yaml
actually only contains a subset of the npm packages in the repository.
This means if a package is not listed in the workspace.yaml it should
be downloaded from the repository and therefore the dependency version
constraint needs to be updated which does not happen at the moment.

To make it work the releng module now uses the same mechanism pnpm
itself uses to resolve workspace modules by using
@pnpm/find-workspace-packages.

Change-Id: I25acb5b36af87d5d8749ff212e12a2e78672d549
Reviewed-on: https://git.eclipse.org/r/160896
Tested-by: CI Bot
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>