Simplify jquery-scout
Reverted to the "old" style which was simpler and easier to maintain
(no need for adding the functions in a separate block to $ and $.fn,
better add them directly). Extending from $ is actually not correct too
because it is a plugin and there may be other plugins (e.g. the
selectors or even third party plugins) and all want to contribute
functions which is not possible using inheritance.

Added typedef for $ so that original jquery functions are available
as well in addition to the jquery-scout functions when using $ as type
in jsdoc.

Changed import syntax for jquery to the syntax recommended by jquery
team (default import). This also makes sure IntelliJ recognizes
$.fn and $.log (at least with 2019.2.3, with 2020.1 it worked with
the other syntax as well).