TileGrid: fix filter animation when tile contains a widget
If a tile has a layout it typically wants to read the current size
during layouting in order to set the size for its children. This size
will probably be wrong if layout is called while an animation runs ->
Delay the animation start until layouting is done.

The fix in graphics.js is necessary for tiles which have the
animate-visible and animate-insert classes at the same time and actually
solves a different problem but belongs to the same topic. It only
affects tiles which have useUiHeight set to true, because otherwise
the logical grid row height is used which is always correct independent
of the animation.

239602

---
Backported from 10.0:
- 4483a5b1930db4b41f1aed4bf38d7252c5ac3359

Change-Id: Ifcfe121f2f057c982a7d8d566656fbcb869bb679
Reviewed-on: https://git.eclipse.org/r/161680
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>
Tested-by: Beat Schwarzentrub <bsh@bsiag.com>