Use new ES6 features: rest param, spread operator, let, arrow function
Code migration:
- let: eslint --fix, some manual fixes
- arrow function: eslint --fix and intellij inspections
- rest param and spread operator: manual fixes

Enabled eslint rules for the new features.

Also adjusted source map generations. With the current devtool type,
'this' inside an arrow function was wrong while debugging.

Also: arrays.pushAll now ensures that the second argument is an array.
Before using the spread operator, the function did nothing
if undefined or an object instead of an array was passed.
Now it would throw an exception. To not break existing code,
the function now auto converts the second argument to an array.